% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_images.R
\name{load_images}
\alias{load_images}
\title{Obtain link to download 1099 leaf images}
\usage{
load_images()
}
\description{
A repository of leaf images of medicinal plants in Sri Lanka. The repository contains 1099 leaf images of 31 species and 29-45 images per species.
}
